Refactor consent-storage.ts into logs.ts
