+++ src/config.ts
@@ -21,11 +21,6 @@ const cmpDomain = (): string =>
         ? 'https://manage.theguardian.com'
         : 'https://manage.code.dev-theguardian.com';
 
-const cmpLogsUrl = (): string =>
-    isProd
-        ? 'https://consent-logs.guardianapis.com/report'
-        : 'https://consent-logs.code.dev-guardianapis.com/report';
-
 const iabVendorListUrl = (): string =>
     isProd
         ? 'https://www.theguardian.com/commercial/cmp/vendorlist.json'
@@ -33,7 +28,6 @@ const iabVendorListUrl = (): string =>
 
 export const CMP_DOMAIN = cmpDomain();
 export const CMP_URL = `${CMP_DOMAIN}/consent`;
-export const CMP_LOGS_URL = cmpLogsUrl();
 export const IAB_VENDOR_LIST_URL = iabVendorListUrl();
 export const IAB_CMP_ID = 112;
 export const IAB_CMP_VERSION = 1;
